add unit test cases

+++ Backend/ResisterColorCodes/ResisterColorCodes/Models/IOhmValueCalculator.cs
@@ -16,6 +16,6 @@ namespace ResisterColorCodes.Models
         /// <param name="bandCColor">The color of the decimal multiplier band.</param>
 
         /// <param name="bandDColor">The color of the tolerance value band.</param>
-        int CalculateOhmValue(string bandAColor, string bandBColor, string bandCColor, string bandDColor);
+        decimal CalculateOhmValue(string bandAColor, string bandBColor, string bandCColor, string bandDColor);
     }
 }
